requirements.txt
pymysql
pyyaml
python-dotenv
